third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This test ensures that script-added import doesn't block scripts");

window.jsTestIsAsync = true;

var importIsLoaded = false;
var scriptIsExecuted = false;

var linkEl = document.createElement("link");
linkEl.setAttribute("rel", "import");
linkEl.setAttribute("href", "resources/async-child.html");
document.head.appendChild(linkEl);
linkEl.onload = function(evt) {
    debug("In onload handler - This should be called later.");
    shouldBeTrue("importIsLoaded");
    shouldBeTrue("scriptIsExecuted");
    finishJSTest();
};

</script>
<script>
debug("In script tag - This should be called first.");
shouldBeTrue("importIsLoaded");
scriptIsExecuted = true;
</script>
</body>
</html>
