third_party/WebKit/LayoutTests/fast/html/imports/import-dynamic-expected.txt

This test ensures that script-added import doesn't block scripts

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


In onload handler - This should be called later.
PASS importIsLoaded is true
FAIL scriptIsExecuted should be true. Was false.
In script tag - This should be called first.
PASS importIsLoaded is true
PASS successfullyParsed is true

TEST COMPLETE
